+++ CHANGELOG.md
@@ -17,6 +17,8 @@
 - cast duration for abilities with a remaining duration of more than 60 seconds will now be hidden to reduce clutter
   - with 12.0.5, this will again render these durations, but properly formatted (e.g. 59m)
   - with 12.0.5, casts with a duration of more than 3 seconds will not show fractions above 3 seconds
+- extended support for LibSharedMedia Borders
+- added support for QUI and Cell
 
 ## Version 1.1.18